requirements.txt
streamlit
torch
torchvision
pillow
opencv-python-headless
numpy
fastapi
uvicorn
python-multipart
